requirements.txt
ExifRead==3.0.0
ipywidgets==7.6.5
lmdb==1.3.0
matplotlib==3.4.3
mmcv==2.0.1
numpy==1.20.3
opencv_python==4.5.5.64
opencv_python_headless==4.5.5.64
pandas==1.3.4
Pillow==10.0.0
plotly==5.15.0
pybind11==2.11.1
pytest==6.2.4
pytorch_msssim==0.2.1
PyYAML==6.0
range_coder==1.1
rawpy==0.18.1
scipy
seaborn==0.11.2
setuptools==58.0.4
tensorboardX==2.6.1
tensorboardX==2.6.1
thop==0.1.1.post2207130030
tifffile==2021.7.2
torch
torchac
torchinterp1d.egg
torchvision
tqdm==4.62.3
scikit_image